requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
SQLAlchemy
pydantic